fused: int64 indexing in _silu_mul_kernel -- fixes a hard crash past ~110k prefill tokens
row * 2 * inter overflows int32 once rows*inter*2 > 2**31, i.e. prefill M = B*S
beyond ~110376 tokens at inter=9728. Prefill ALWAYS takes this path (M = B*S
always exceeds BM_MAX), so any workload past that size kills the entire run,
not just that workload.

Confirmed as a hard CUDA illegal-memory-access at B256 x 512 and B64 x 2048
(both M=131072); B128 x 512 (M=65536) is clean. Found during the large-batch
audit. Widening the index changes no arithmetic and no rounding point;
selftest max logit diff is unaffected.

Every official run to date passes, so the current hidden set stays under the
threshold -- this is removing a landmine, not fixing an observed failure.

+++ engine/fused.py
@@ -152,7 +152,11 @@ def _add_rms_kernel(
 
 @triton.jit
 def _silu_mul_kernel(gu_ptr, out_ptr, n, inter, BLOCK: tl.constexpr):
-    idx = tl.program_id(0) * BLOCK + tl.arange(0, BLOCK)
+    # int64 indexing: row * 2 * inter overflows int32 once rows*inter*2 > 2**31, i.e. prefill
+    # M = B*S beyond ~110k tokens at inter=9728. Confirmed as an illegal-memory-access crash at
+    # B256 S512 and B64 S2048 (both M=131072); prefill always takes this path, so any workload
+    # past that size kills the whole run. Widening the index changes no arithmetic or rounding.
+    idx = tl.program_id(0).to(tl.int64) * BLOCK + tl.arange(0, BLOCK).to(tl.int64)
     m = idx < n
     row = idx // inter
     col = idx % inter
